refactor(ui): split node-graph and diff selection into feature folders
Keep packages/ui/src and diff-view under directory-density hard caps after new helpers.
