+++ packages/ui/src/diff-view/selection/copy-sticky.ts
