+++ src/renderer/lib/context-menu/use-menu.ts
@@ -1,7 +1,7 @@
 import {
   getDiffCopyStickyText,
   isDiffCopyStickySurface,
-} from "@pier/ui/diff-view/copy-sticky.ts";
+} from "@pier/ui/diff-view/selection/copy-sticky.ts";
 import {
   releaseTooltipSuppression,
   suppressTooltips,
